Upgrade swagger-parser to fix vulnerability
Fixes Vulnerability with dependency swagger-parser v10.0.2 #287.

swagger-parser depends on z-schema, which depends on validator, and the
version that swagger-parser v10.0.2 depends on has a security
vulnerability (https://github.com/advisories/GHSA-xx4c-jj58-r7x6).

PR #300 tried to fix this, but because the version is pinned to 10.0.2
exactly in package.json it hasn't changed the version used.